Rename to arrow_string
