+++ crates/re_types/src/lib.rs
@@ -502,8 +502,8 @@ mod size_bytes;
 pub use component_name::ComponentName;
 pub use size_bytes::SizeBytes;
 
-mod arrow_adapter;
-pub use arrow_adapter::ArrowString;
+mod arrow_string;
+pub use arrow_string::ArrowString;
 
 #[cfg(feature = "testing")]
 pub mod testing;
